Create follow API, JS functions not working yet
When I am creating the get request in the URL field, everything is working.
When I am trying to attach this functionality to the button it is not
working, yet, throwing "Uncaught (in promise) SyntaxError:
JSON.parse: unexpected character at line 1 column 1 of the JSON data"

Also I need to make the button change it's value when somebody
clicked it.